Set the maximum load factor when building the perfect hash function
We used to build the perfect hash function with the CHD_PH algorithm
with the default load factor value which was 0.5. We thus ended up
a hash function which itself took the minimum amount of space, but
would result in an index twice as large as the number of
objects in the shard.

We now set the load factor to the maximum value (0.99) using the
ill-named `cmph_config_set_graphsize()` function to loose the
minimum amount of space in the index.

The 1% lost will still amout to 11-12 MiB for a 30 millions objects
shard.