extension: initialize index entries as "deleted" entries
A deleted entry is stored using \x00<x32> as key and UINT64_MAX as
value. Initialize all the entries in this index with this special value
to make it clearer these are unset entries (since the computed CMPH is
not really *minimal*, the general case will have a few empty index entries,
so make it explicit these are empty using the special "deleted" value).