Migrate to pybind11 and restructure the source code directory
Use pybind11 to wrap the cmph and shard manipulation code instead of
cffi, it makes is a bit easier to add (C/C++) features in the extension.

The extension source files have been moved to src/_shard, and the python
source files for the swh.shard package have been moved to src/swh/shard,
moving away from all other swh package structure. This is required to
prevent side effects of having the local 'swh' directory in the working
directory of the developer, thus in the sys.path (by default), breaking
the dark magic involved in the loading of the package when it is
installed in editable mode (i.e. not to break pytest when executed
directly from the source with the package being installed in editable
mode).