tests: Fix a flakiness behavior of test_cli
Issue was the shard_index_get() (used in ShardReader.getindex()) returns
the index entry at given position, but this can be NULL, while the test
wrongly made the assumption it would always return a valid index.

Fix the test, but also add a new sahrd_cmph_search() C function and a
corresponding getpos() method to the ShardReader class to make it
possible to retrieve the index position from the key in python (the
API did not actually permitted to access the result of a call to
cmph_search()).